backend/pipeline/scrapers/facebookScriping/faccebook_sidwaya/requirements.txt
requests==2.31.0
playwright==1.48.0
python-dotenv==1.0.0
